@@ -1,43 +1,39 @@
 cmake_minimum_required(VERSION 3.1)
-cmake_policy(SET CMP0048 NEW)
-project( EyesThatBlink VERSION 1.0.1)
+project(EyesThatBlink VERSION 1.0.1)
 
-add_definitions( --std=c++11 )
+cmake_policy(SET CMP0048 NEW)
+
+add_definitions(--std=c++14)
 
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_CURRENT_SOURCE_DIR}/cmake)
 include_directories( ${CMAKE_CURRENT_SOURCE_DIR}/external/plog/include )
 include_directories( ${CMAKE_CURRENT_SOURCE_DIR}/external/pstreams-1.0.1/)
 
-# set(USE_BOOST_PROCESS OFF)
-# if(USE_BOOST_PROCESS)
-#     add_definitions( -DUSE_BOOST_PROCESS )
-#     include_directories( ${CMAKE_CURRENT_SOURCE_DIR}/external/cpp-subprocess/)
-# endif( )
-
 find_package(OpenCV REQUIRED)
 
-include_directories( ${OpenCV_INCLUDE_DIRS} )
-message( STATUS "OpenCV libraries ${OpenCV_LIBRARIES}" )
+include_directories(${OpenCV_INCLUDE_DIRS})
+message(STATUS "OpenCV libraries ${OpenCV_LIBRARIES}")
 
 # Check for pkg-config modules. We need to link with their libs.
-find_package( PkgConfig REQUIRED )
+find_package(PkgConfig REQUIRED)
 if(UNIX)
     pkg_check_modules(LIBNOTIFY libnotify REQUIRED)
     list(APPEND OTHER_LIBS ${LIBNOTIFY_LIBRARIES})
     pkg_check_modules(GTKMM gtkmm-3.0 REQUIRED)
     include_directories(${GTKMM_INCLUDE_DIRS})
-endif( )
+endif()
 
-# pthreads
-find_package( Threads )
+find_package(Threads)
+find_package(Boost
+    COMPONENTS program_options regex date_time iostreams filesystem system
+    REQUIRED)
+
 list(APPEND OTHER_LIBS ${CMAKE_THREAD_LIBS_INIT} )
 
-find_package( Boost REQUIRED COMPONENTS iostreams system filesystem date_time
-    regex)
-include_directories( ${Boost_INCLUDE_DIRS} )
-include_directories( ${CMAKE_CURRENT_SOURCE_DIR}/external/boost-process/include )
-include_directories( ${CMAKE_CURRENT_SOURCE_DIR}/external/boost-asio/include )
-include_directories( ${CMAKE_CURRENT_SOURCE_DIR}/external/boost-winapi/include )
+include_directories(${Boost_INCLUDE_DIRS})
+include_directories(${CMAKE_CURRENT_SOURCE_DIR}/external/boost-process/include)
+include_directories(${CMAKE_CURRENT_SOURCE_DIR}/external/boost-asio/include)
+include_directories(${CMAKE_CURRENT_SOURCE_DIR}/external/boost-winapi/include)
 
 if(WIN32)
     add_definitions( -DOS_IS_WINDOWS )
