@@ -1,55 +1,61 @@
 /**
-*       Filename:  main.cpp
-*
-*    Description:  Entry point of the application.
-*
-*        Version:  0.0.1
-*        Created:  2017-04-16
-*       Revision:  none
-*
-*         Author:  Dilawar Singh <dilawars@ncbs.res.in>
-*   Organization:  NCBS Bangalore
-*
+ *       Filename:  main.cpp
+ *
+ *    Description:  Entry point of the application.
+ *
+ *        Version:  0.0.1
+ *        Created:  2017-04-16
+ *       Revision:  none
+ *
+ *         Author:  Dilawar Singh <dilawars@ncbs.res.in>
+ *   Organization:  NCBS Bangalore
+ *
  *        License:  GNU GPL2
  */
 
-
 #include <iostream>
-
-#include "core/globals.h"
-#include "core/main_loop.h"
+#include <boost/filesystem.hpp>
+#include <boost/program_options.hpp>
 #include "plog/Log.h"
+
 #include "config.h"
+#include "core/globals.h"
 #include "core/helpers.h"
+#include "core/main_loop.h"
 #include "ui/ui_unix.h"
 
-#include <boost/filesystem.hpp>
-#include <plog/Appenders/ConsoleAppender.h>
 
 using namespace std;
+
+namespace po = boost::program_options;
+
+extern boost::program_options::variables_map vm_;
 
 /**
  * @function main
  */
-int main( int argc, char* argv[] )
+int main(int argc, char* argv[])
 {
-    // No log should be written to stdout.
-    boost::filesystem::path logpath( expand_user( LOG_FILE_PATH ) );
-    static plog::RollingFileAppender<plog::CsvFormatter> 
-        fileAppender( logpath.c_str( ), 8000, 3 );
+    // Declare the supported options.
+    po::options_description desc("Allowed options");
+    desc.add_options()
+        ("help,h", "produce help message")
+        ("configfile,c", po::value<string>(), "filepath of config file")
+        ("datadir,d", po::value<string>(), "directory containing application data.")
+        ;
 
-#ifndef NDEBUG
-    static plog::ConsoleAppender<plog::TxtFormatter> consoleAppender;
-    plog::init(plog::debug, &consoleAppender);
-#endif
+    po::store(po::parse_command_line(argc, argv, desc), vm_);
+    po::notify(vm_);
 
-    LOG_INFO << "Initializing ";
+    plog::init(plog::debug);
 
-    init_camera( );
+    if(vm_.count("help")) {
+        cout << desc << endl;
+        return 1;
+    }
 
-    unix_ui( argc, argv );
-
+    init_camera();
+    unix_ui(argc, argv);
     return 0;
 }
-
 
